backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
httpx==0.25.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
python-telegram-bot==20.3
apscheduler==3.10.4
chromadb==0.4.16
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
jwcrypto==1.5.4
python-multipart==0.0.6
cors==1.0.1
psycopg2-binary==2.9.9
